move: interface 파일 분리
